Add fonts, player health, ui in-game
