+++ Assets/Fonts/ButtonFont SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7501af3effad528429b5d92ff71e536f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
