+++ Assets/Fonts/TituloFont SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5bf8a081682dc3046907e3fb6e30e56b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
