+++ Assets/Prefabs/Heart.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1372ce54243a4a4d8a5d3ec596687e8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
