--- Assets/Scripts/Enemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: acce4d713e51d4348ad9354a3c76667c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
