--- Assets/Scripts/EnemySpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9eb4a5a70d257284e8642c9de68cd2e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
