--- Assets/Scripts/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b0901edfe78e4a41a703d58b6891460
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
