+++ Assets/Scripts/Player/Player.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 1b0901edfe78e4a41a703d58b6891460
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - bullet: {fileID: 184337399719150429, guid: eddc076e9cc18ad44948fec38acfab26, type: 3}
+  - spawnerBulletPos: {instanceID: 0}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
