+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - MoneyUI
   - Gun
   - bullet
+  - LifeUI
   layers:
   - Default
   - TransparentFX
